+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2022-02-08
+
+### Added
+
+* Added iOS implementation of method `selectPoi(...)` to select a POI in the map.
+* Added iOS implementation of method `navigateToPoi(...)` to request navigation to a POI.
+
 ## [0.0.9] - 2022-01-28
 
 # Added